강한 커플링 문제 해결 - 메소드 setModel()로 수정
