+++ DailyDiary/detailViewController.swift
@@ -1,0 +1,25 @@
+//
+//  detailView.swift
+//  DailyDiary
+//
+//  Created by sutie on 2017. 12. 27..
+//  Copyright © 2017년 sutie. All rights reserved.
+//
+
+import UIKit
+
+class detailViewController: UIViewController {
+
+    @IBOutlet weak var genderLabel: UILabel!
+    
+    @IBOutlet weak var locationLabel: UILabel!
+    
+    var details: [ProfileList] = []
+    
+   // genderLabel.text = details[IndexPath.row].gender
+  //  locationLabel.text = details[IndexPath.row].location
+    
+    
+    
+    
+}
